requirements.txt
pandas==2.3.0
jupyter==1.1.1
numpy==2.3.1
argparse==1.4.0
matplotlib==3.10.3
tqdm==4.67.1
openai==1.93.0
pathlib==1.0.1
python-dotenv==1.1.1